feature(rank-api): Move some endpoints to the users tag
