+++ api/src/routes/users.ts
@@ -6,11 +6,46 @@
  */
 
 import express, { Request, Response } from "express";
-import logger from "../utilities/logger";
+import logger from "../utils/logger";
 import bcrypt from "bcrypt";
 import pool from "../database/index";
+import { Rank } from "../database/models";
+import { indexOf } from "lodash";
+import { paginationIsValid } from "../utils";
 
 const router = express.Router();
+
+const getUserConnection = async (userId: string) => {
+  return await pool.query(
+    "SELECT user1_id, user2_id FROM user_connections WHERE user1_id = $1 OR user2_id = $1",
+    [userId]
+  );
+};
+
+const getGameRanks = async (id: number) => {
+  try {
+    const result = await pool.query("SELECT * FROM games WHERE game_id = $1", [
+      id,
+    ]);
+    return result.rows[0].rank_types;
+  } catch (error) {
+    logger.error(error);
+  }
+};
+
+const getAverageRank = async (ranks: Rank[]) => {
+  let average = 0;
+  if (ranks[0].numeric_rank !== null) {
+    ranks.forEach((rank) => (average += rank.numeric_rank));
+    return average / ranks.length;
+  } else {
+    const gameRanks = await getGameRanks(ranks[0].game_id);
+    ranks.forEach((rank) => {
+      average += indexOf(gameRanks, rank.text_rank);
+    });
+    return gameRanks[average / ranks.length];
+  }
+};
 
 /**
  * @swagger
@@ -172,6 +207,7 @@ router.get("/:user_id/connections", async (req: Request, res: Response) => {
       return res.status(404).json({ error: "User not found" });
     }
 
+    // TODO: Update this to use the existing function or update existing function to use this.
     // Search for potential connections based on the search query
     const potentialConnections = await pool.query(
       `SELECT user_id, username, full_name, email, profile_picture_url
@@ -339,11 +375,9 @@ router.post("/:user_id/games/:game_id", async (req: Request, res: Response) => {
     );
 
     if (existingRank.rows.length > 0) {
-      return res
-        .status(400)
-        .json({
-          error: "User already has a rank for today and the specified game",
-        });
+      return res.status(400).json({
+        error: "User already has a rank for today and the specified game",
+      });
     }
 
     // Insert the user's rank for today and the specified game
@@ -358,5 +392,235 @@ router.post("/:user_id/games/:game_id", async (req: Request, res: Response) => {
     res.status(500).json({ error: "Internal Server Error" });
   }
 });
+
+// Rank endpoints
+
+/**
+ * @swagger
+ * /users/{userId}/games/{gameId}/ranks:
+ *   get:
+ *     summary: Get all ranks of a user on a game
+ *     tags: [Users]
+ *     description: Retrieve all ranks for a specific user on a specific game from the database.
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         description: ID of the user to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *       - in: path
+ *         name: gameId
+ *         required: true
+ *         description: ID of the game to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *     responses:
+ *       200:
+ *         description: Successful response
+ *       500:
+ *         description: Internal Server Error
+ */
+
+router.get(
+  "/:userId/games/:gameId/ranks",
+  async (req: Request, res: Response) => {
+    try {
+      const { userId, gameId } = req.params;
+
+      const result = await pool.query(
+        "SELECT * FROM user_ranks WHERE user_id = $1 AND game_id = $2",
+        [userId, gameId]
+      );
+      res.json(result.rows);
+    } catch (error) {
+      logger.error(error);
+      res.status(500).send("Internal Server Error");
+    }
+  }
+);
+
+/**
+ * @swagger
+ * /users/{userId}/ranks:
+ *   get:
+ *     summary: Get rank by user id
+ *     tags: [Users]
+ *     description: Retrieve all ranks for a specific user from the database.
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         description: ID of the user to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *       - in: query
+ *         name: page
+ *         description: Page number for pagination
+ *         schema:
+ *           type: integer
+ *         default: 1
+ *       - in: query
+ *         name: pageSize
+ *         description: Number of logs per page
+ *         schema:
+ *           type: integer
+ *         default: 10
+ *     responses:
+ *       200:
+ *         description: Successful response
+ *       500:
+ *         description: Internal Server Error
+ */
+
+router.get("/:userId/ranks", async (req: Request, res: Response) => {
+  try {
+    const { userId } = req.params;
+    const { page = 1, pageSize = 10 } = req.query;
+
+    if (!paginationIsValid(page as number, pageSize as number)) {
+      return res
+        .status(400)
+        .json({ error: "Invalid page or pageSize parameters" });
+    }
+
+    // Calculate the offset for pagination
+    const offset = (Number(page) - 1) * Number(pageSize);
+    const result = await pool.query(
+      "SELECT * FROM user_ranks WHERE user_id = $1 ORDER BY rank_date DESC LIMIT $2 OFFSET $3",
+      [userId, Number(pageSize), offset]
+    );
+    res.json(result.rows);
+  } catch (error) {
+    logger.error(error);
+    res.status(500).send("Internal Server Error");
+  }
+});
+
+/**
+ * @swagger
+ * /users/{userId}/connections/games/{gameId}/ranks:
+ *   get:
+ *     summary: Get all ranks of a user and their connections on a game
+ *     tags: [Users]
+ *     description: Retrieve all ranks of a user and their connections on a specific game from the database.
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         description: ID of the user to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *       - in: path
+ *         name: gameId
+ *         required: true
+ *         description: ID of the game to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *       - in: query
+ *         name: page
+ *         description: Page number for pagination
+ *         schema:
+ *           type: integer
+ *         default: 1
+ *       - in: query
+ *         name: pageSize
+ *         description: Number of logs per page
+ *         schema:
+ *           type: integer
+ *         default: 10
+ *     responses:
+ *       200:
+ *         description: Successful response
+ *       500:
+ *         description: Internal Server Error
+ */
+
+// TODO: Add a since/until parameter.
+router.get(
+  "/:userId/connections/games/:gameId/ranks",
+  async (req: Request, res: Response) => {
+    try {
+      const { userId, gameId } = req.params;
+      const { page = 1, pageSize = 10 } = req.query;
+
+      if (!paginationIsValid(page as number, pageSize as number)) {
+        return res
+          .status(400)
+          .json({ error: "Invalid page or pageSize parameters" });
+      }
+
+      // Calculate the offset for pagination
+      const offset = (Number(page) - 1) * Number(pageSize);
+
+      const userConnections = await getUserConnection(userId);
+
+      let connectionIds = userConnections.rows.map((connection) => [
+        connection.user1_id,
+        connection.user2_id,
+      ]);
+
+      let stringedConnectionIds = [...new Set(connectionIds.flat())].join(",");
+
+      const result = await pool.query(
+        `SELECT * FROM user_ranks WHERE user_id IN (${stringedConnectionIds}) AND game_id = $1 ORDER BY rank_date DESC LIMIT $2 OFFSET $3`,
+        [gameId, Number(pageSize), offset]
+      );
+
+      res.json(result.rows);
+    } catch (error) {
+      logger.error(error);
+      res.status(500).send("Internal Server Error");
+    }
+  }
+);
+
+/**
+ * @swagger
+ * /users/{userId}/games/{gameId}/ranks/average:
+ *   get:
+ *     summary: Get an average of the ranks of a user on a game
+ *     tags: [Users]
+ *     description: Retrieve the average rank of a user on a specific game from the database.
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         description: ID of the user to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *       - in: path
+ *         name: gameId
+ *         required: true
+ *         description: ID of the game to retrieve ranks for
+ *         schema:
+ *           type: integer
+ *     responses:
+ *       200:
+ *         description: Successful response
+ *       500:
+ *         description: Internal Server Error
+ */
+
+// TODO: Add a since/until parameter.
+router.get(
+  "/:userId/games/:gameId/ranks/average",
+  async (req: Request, res: Response) => {
+    try {
+      const { userId, gameId } = req.params;
+
+      const result = await pool.query(
+        "SELECT * FROM user_ranks WHERE user_id = $1 AND game_id = $2",
+        [userId, gameId]
+      );
+
+      const ranks = result.rows;
+      res.json(await getAverageRank(ranks));
+    } catch (error) {
+      logger.error(error);
+      res.status(500).send("Internal Server Error");
+    }
+  }
+);
 
 export default router;
